Boatrental merged + test added
