+++ src/main/java/nl/yacht/lakesideresort/Main.java
@@ -1,10 +1,5 @@
 package nl.yacht.lakesideresort;
 
-import nl.yacht.lakesideresort.domain.BoatRental;
-import nl.yacht.lakesideresort.domain.LakeTrip;
-import nl.yacht.lakesideresort.domain.RiverTrip;
-import nl.yacht.lakesideresort.domain.Trip;
-
 public class Main {
     public static void main(String[] args) {
